refactor: use a disposable instead of try-finally
Created a TraceListenerManager that optionally adds the
DefaultTraceListener to the system's list of Trace listeners. The
manager is disposable and uses its Dispose() behavior to remove the
listener if one was added. By putting the manager in a using, we
obtain reliable removal of the DefaultTraceListener without using a
try-finally block.