fix: don't mark failing tests passed and failed
When a [Failure] method exists in the test suite, failing methods are
only marked failed. The test executor correctly reports passed/failed
status for the [Failure] method when executing it, instead of the test
method that triggered the failure.

Previously, failing methods in a suite with a [Failure] method were
marked passed after being marked failed, and status for the [Failure]
method was not reported.